pyproject.toml
[project]
name = "servicebus-mcp"
version = "0.1.0"
description = "MCP server for sending and inspecting messages on Azure Service Bus"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0",
    "azure-servicebus>=7.12",
    "azure-identity>=1.17",
]

[project.scripts]
servicebus-mcp = "servicebus_mcp.server:main"

[tool.uv]
package = true

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/servicebus_mcp"]
